CLOUD-42118 Split AmbariClient to services
