+++ build.gradle
@@ -53,7 +53,7 @@ repositories {
 dependencies {
     compile 'org.slf4j:slf4j-api:1.7.7'
     compile 'org.slf4j:slf4j-log4j12:1.7.7'
-    compile 'org.codehaus.groovy:groovy-all:2.1.8'
+    compile 'org.codehaus.groovy:groovy-all:2.4.4'
     compile 'org.codehaus.groovy.modules.http-builder:http-builder:0.7.1'
     compile 'commons-io:commons-io:2.4'
     compile "org.apache.httpcomponents:httpcore:4.3.1"
