+++ src/main/groovy/com/sequenceiq/ambari/client/AmbariClient.groovy
@@ -16,45 +16,35 @@
  * limitations under the License.
  */
 package com.sequenceiq.ambari.client
+
+import com.sequenceiq.ambari.client.services.AlertService
+import com.sequenceiq.ambari.client.services.BlueprintService
+import com.sequenceiq.ambari.client.services.ConfigService
+import com.sequenceiq.ambari.client.services.HBaseService
+import com.sequenceiq.ambari.client.services.ServiceAndHostService
+import com.sequenceiq.ambari.client.services.KerberosService
+import com.sequenceiq.ambari.client.services.TaskService
+import com.sequenceiq.ambari.client.services.UserService
+import com.sequenceiq.ambari.client.services.StackService
 import groovy.json.JsonBuilder
 import groovy.json.JsonSlurper
-import groovy.text.SimpleTemplateEngine
 import groovy.util.logging.Slf4j
 import groovyx.net.http.ContentType
 import groovyx.net.http.HttpResponseException
 import groovyx.net.http.RESTClient
-import org.apache.commons.io.IOUtils
-import org.apache.http.NoHttpResponseException
-import org.apache.http.client.ClientProtocolException
-import org.apache.http.config.Registry
-import org.apache.http.config.RegistryBuilder
-import org.apache.http.conn.socket.ConnectionSocketFactory
-import org.apache.http.conn.socket.PlainConnectionSocketFactory
-import org.apache.http.conn.ssl.SSLConnectionSocketFactory
-import org.apache.http.conn.ssl.SSLContexts
 import org.apache.http.impl.client.HttpClientBuilder
 import org.apache.http.impl.conn.PoolingHttpClientConnectionManager
 
 import javax.net.ssl.SSLContext
-import java.security.Security
-import java.net.ConnectException
-import java.net.NoRouteToHostException
-import java.net.UnknownHostException
-
 /**
  * Basic client to send requests to the Ambari server.
  */
 @Slf4j
-class AmbariClient {
+class AmbariClient implements AlertService, BlueprintService, ConfigService, HBaseService, ServiceAndHostService, KerberosService, StackService, TaskService, UserService {
 
-  private static final int PAD = 30
-  private static final int OK_RESPONSE = 200
-  private static final String SLAVE = "slave_"
-  boolean debugEnabled = false;
-  def RESTClient ambari
-  def slurper = new JsonSlurper()
-  def templateProcessor = new SimpleTemplateEngine()
-  def clusterName
+  private static final String SLAVE = 'slave_'
+
+  final RESTClient ambari
 
   /**
    * Connects to the ambari server.
@@ -67,15 +57,15 @@ class AmbariClient {
    * @param clientKeyPath client key path, used in 2-way-ssl
    * @param serverCertPath server certificate path, used in 2-way-ssl
    */
-  AmbariClient(host = 'localhost', port = '8080', user = 'admin', password = 'admin',
-               clientCertPath = null, clientKeyPath = null, serverCertPath = null) {
-    def http = clientCertPath == null ? "http" : "https";
+  AmbariClient(String host = 'localhost', String port = '8080', String user = 'admin', String password = 'admin',
+               String clientCertPath = null, String clientKeyPath = null, String serverCertPath = null) {
+    def http = clientCertPath == null ? 'http' : 'https';
     ambari = new RESTClient("$http://${host}:${port}/api/v1/" as String)
 
     if (clientCertPath) {
-      SSLContext sslContext = setupSSLContext(clientCertPath, clientKeyPath, serverCertPath);
+      SSLContext sslContext = utils.setupSSLContext(clientCertPath, clientKeyPath, serverCertPath);
       PoolingHttpClientConnectionManager connectionManager =
-        new PoolingHttpClientConnectionManager(setupSchemeRegistry(sslContext));
+        new PoolingHttpClientConnectionManager(utils.setupSchemeRegistry(sslContext));
       connectionManager.setMaxTotal(1000);
       connectionManager.setDefaultMaxPerRoute(500);
       def httpClient = HttpClientBuilder.create().setConnectionManager(connectionManager)
@@ -99,378 +89,11 @@ class AmbariClient {
   }
 
   /**
-   * Sets the debug variable. Used by printing the API calls for the Ambari Shell.
-   *
-   * @param enabled enable or disable
-   */
-  def setDebugEnabled(boolean enabled) {
-    debugEnabled = enabled;
-  }
-
-  /**
-   * Returns the name of the cluster.
-   *
-   * @return the name of the cluster of null if no cluster yet
-   */
-  def String getClusterName() {
-    if (!clusterName) {
-      def clusters = getClusters();
-      if (clusters) {
-        clusterName = clusters.items[0]?.Clusters?.cluster_name
-      }
-    }
-    return clusterName
-  }
-
-  /**
-   * Adds the registered nodes to the cluster.
-   *
-   * @param hosts list of hostname to add
-   * @throws HttpResponseException if a node is not registered with ambari
-   */
-  def addHosts(List<String> hosts) throws HttpResponseException {
-    def requestBody = hosts.collectAll { ["Hosts": ["host_name": it]] }
-    ambari.post(path: "clusters/${getClusterName()}/hosts", body: new JsonBuilder(requestBody).toPrettyString(), { it })
-  }
-
-  /**
-   * Returns the state of the host.
-   *
-   * @param host host's internal hostname
-   * @return state of the host
-   */
-  String getHostState(String host) {
-    getAllResources("hosts/$host").Hosts.host_status
-  }
-
-  /**
-   * Create a new Ambari user.
-   */
-  def createUser(String user, String password, boolean admin) {
-    def context = ["Users/active": true, "Users/admin": admin, "Users/password": password, "Users/user_name": user]
-    ambari.post(path: "users", body: new JsonBuilder(context).toPrettyString(), { it })
-  }
-
-  /**
-   * Delete an Ambari user.
-   */
-  def deleteUser(String user) {
-    ambari.delete(path: "users/$user")
-  }
-
-  /**
-   * Change the password of an Ambari user.
-   */
-  def changePassword(String user, String oldPassword, String newPassword, boolean admin) {
-    def roles = ["user"]
-    if (admin) {
-      roles << "admin"
-    }
-    def context = ["Users/password": newPassword, "Users/old_password": oldPassword]
-    ambari.put(path: "users/$user", body: new JsonBuilder(context).toPrettyString(), requestContentType: ContentType.URLENC)
-  }
-
-  /**
    * Runs a MapReduce service check which is a simple WordCount.
    * @return id of the request
    */
   def int runMRServiceCheck() {
-    runServiceCheck("MAPREDUCE2_SERVICE_CHECK", "MAPREDUCE2")
-  }
-
-  /**
-   * Run a service check.
-   *
-   * @param command command to run
-   * @param serviceName name of the service
-   * @return id of the request
-   */
-  def int runServiceCheck(String command, String serviceName) {
-    Map bodyMap = [
-      "RequestInfo"              : [command: command, context: command],
-      "Requests/resource_filters": [[service_name: serviceName]]
-    ]
-    ambari.post(path: "clusters/${getClusterName()}/requests", body: new JsonBuilder(bodyMap).toPrettyString(), {
-      getRequestId(it)
-    })
-  }
-
-  /**
-   * Decommission and remove a host from the cluster.
-   * NOTE: this is a synchronous call, it wont return until all
-   * requests are finished
-   *
-   * Steps:
-   *  1, decommission services
-   *  2, stop services
-   *  3, delete host components
-   *  4, delete host
-   *  5, restart services
-   *
-   * @param hostName host to be deleted
-   */
-  def removeHost(String hostName) {
-    def components = getHostComponentsMap(hostName).keySet() as List
-
-    // decommission
-    if (components.contains("NODEMANAGER")) {
-      decommissionNodeManagers([hostName])
-    }
-    if (components.contains("DATANODE")) {
-      decommissionDataNodes([hostName])
-    }
-
-    // stop services
-    def requests = stopComponentsOnHost(hostName, components)
-    waitForRequestsToFinish(requests.values() as List)
-
-    // delete host components
-    deleteHostComponents(hostName, components)
-
-    // delete host
-    deleteHost(hostName)
-
-    // restart zookeper
-    def id = restartServiceComponents("ZOOKEEPER", ["ZOOKEEPER_SERVER"])
-    waitForRequestsToFinish([id])
-
-    // restart nagios
-    if (getServiceComponentsMap().containsKey("NAGIOS")) {
-      id = restartServiceComponents("NAGIOS", ["NAGIOS_SERVER"])
-      waitForRequestsToFinish([id])
-    }
-  }
-
-  /**
-   * Does not return until all the requests are finished.
-   * @param requestIds ids of the requests
-   */
-  def waitForRequestsToFinish(List<Integer> requestIds) {
-    def stopped = false
-    while (!stopped) {
-      def state = true
-      for (int id : requestIds) {
-        if (getRequestProgress(id) != 100.0) {
-          state = false;
-          break;
-        }
-      }
-      stopped = state
-      Thread.sleep(2000)
-    }
-  }
-
-  /**
-   * Decommission the data node on the given hosts.
-   */
-  def int decommissionDataNodes(List<String> hosts) {
-    decommission(hosts, "DATANODE", "HDFS", "NAMENODE")
-  }
-
-  /**
-   * Decommission the node manager on the given hosts.
-   */
-  def int decommissionNodeManagers(List<String> hosts) {
-    decommission(hosts, "NODEMANAGER", "YARN", "RESOURCEMANAGER")
-  }
-
-  /**
-   * Decommission the HBase Region Server on the given hosts.
-   */
-  def int decommissionHBaseRegionServers(List<String> hosts) {
-    decommission(hosts, "HBASE_REGIONSERVER", "HBASE", "HBASE_MASTER")
-  }
-
-  /**
-   * Set the HBase Region servers to maintenance mode and back.
-   *
-   * @param hostNames names of the hosts which is running the Region server
-   * @param mode true to maintenance mode false to active mode
-   */
-  def void setHBaseRegionServersToMaintenance(List<String> hostNames, boolean mode) {
-    hostNames.each {
-      setHBaseRegionServerToMaintenance(it, mode)
-    }
-  }
-
-  /**
-   * Set the HBase Region server to maintenance mode and back.
-   *
-   * @param hostName name of the host which is running the Region server
-   * @param mode true to maintenance mode false to active mode
-   */
-  def void setHBaseRegionServerToMaintenance(String hostName, boolean mode) {
-    def reqInfo = ["RequestInfo": ["context": "Turn On Maintenance Mode for RegionServer"],
-                   "Body"       : ["HostRoles": ["maintenance_state": mode ? "ON" : "OFF"]]]
-    def Map<String, ?> putRequestMap = [:]
-    putRequestMap.put('requestContentType', ContentType.URLENC)
-    putRequestMap.put('path', "clusters/${getClusterName()}/hosts/$hostName/host_components/HBASE_REGIONSERVER")
-    putRequestMap.put('body', new JsonBuilder(reqInfo).toPrettyString());
-
-    ambari.put(putRequestMap)
-  }
-
-  /**
-   * Decommission a host component on a given host.
-   * @param host hostName where the component is installed to
-   * @param slaveName slave to be decommissioned
-   * @param serviceName where the slave belongs to
-   * @param componentName where the slave belongs to
-   */
-  def int decommission(List<String> hosts, String slaveName, String serviceName, String componentName) {
-    def requestInfo = [
-      command   : "DECOMMISSION",
-      context   : "Decommission $slaveName",
-      parameters: ["slave_type": slaveName, "excluded_hosts": hosts.join(',')]
-    ]
-    def filter = [
-      ["service_name": serviceName, "component_name": componentName]
-    ]
-    Map bodyMap = [
-      "RequestInfo"              : requestInfo,
-      "Requests/resource_filters": filter
-    ]
-    ambari.post(path: "clusters/${getClusterName()}/requests", body: new JsonBuilder(bodyMap).toPrettyString(), {
-      getRequestId(it)
-    })
-  }
-
-  /**
-   * Returns the name of the node and the size of the replicated block that are remaining.
-   *
-   * @return Map where the key is the internal host name of the node and the value
-   *         is size of the replicated block
-   */
-  def Map<String, Long> getDecommissioningDataNodes() {
-    def result = [:]
-    def response = slurp("clusters/${getClusterName()}/services/HDFS/components/NAMENODE", "metrics/dfs/namenode/DecomNodes")
-    def nodes = slurper.parseText(response?.metrics?.dfs?.namenode?.DecomNodes)
-    if (nodes) {
-      nodes.each {
-        result << [(it.key): it.value.underReplicatedBlocks as Long]
-      }
-    }
-    result
-  }
-
-  /**
-   * Returns the hostnames of those Region servers which currently stores data and the number of files they store
-   *
-   * @param hostFilter filter for which hosts are interested
-   * @return map key - hostname value - stored files
-   */
-  def Map<String, Long> getHBaseRegionServersWithData(List<String> hostFilter) {
-    def res = [:]
-    def prePredicate = ["fields": "host_components/metrics/hbase"]
-    def host_components = getAllPredictedResources("services/HBASE/components/HBASE_REGIONSERVER", prePredicate)?.host_components
-    if (host_components) {
-      host_components.each {
-        def hostName = it.HostRoles.host_name
-        if (hostFilter.isEmpty() || hostFilter.contains(hostName)) {
-          def storeFiles = it.metrics.hbase.regionserver.storefiles
-          if (storeFiles > 0) {
-            res << ["$hostName": storeFiles]
-          }
-        }
-      }
-    }
-    res
-  }
-
-  /**
-   * Creates a new alert definition.
-   *
-   * @param definition alert definition as json
-   */
-  def void createAlert(String definition) {
-    ambari.post(path: "clusters/${getClusterName()}/alert_definitions", body: definition, { it })
-  }
-
-  /**
-   * Returns the alert definitions.
-   *
-   * @return collection of definitions
-   */
-  def List<Map<String, String>> getAlertDefinitions() {
-    getAllResources("alert_definitions", "AlertDefinition").items.collect {
-      def details = [:]
-      def definition = it.AlertDefinition
-      details << ["enabled": definition.enabled]
-      details << ["scope": definition.scope]
-      details << ["interval": definition.interval as String]
-      details << ["description": definition.description]
-      details << ["name": definition.name]
-      details << ["label": definition.label]
-      details << ["service_name": definition.service_name]
-      details
-    }
-  }
-
-  /**
-   * Returns all the defined alerts grouped by alert definition name.
-   * A filter can be applied for scopes: HOST, ANY, SERVICE
-   *
-   * @return alert properties
-   */
-  def Map<String, List<Map<String, Object>>> getAlerts(List<String> scope = []) {
-    getAllResources("alerts", "Alert").items.findAll {
-      scope == [] || scope.contains(it.Alert.scope)
-    }.collect { it.Alert }.groupBy { it.definition_name }
-  }
-
-  /**
-   * Get a specified alert by id.
-   *
-   * @param id id of the alert
-   * @return alert or empty collection
-   */
-  Map<String, Object> getAlert(int id) {
-    def response = getAllResources("alerts/$id")
-    response ? response.Alert : [:]
-  }
-
-  /**
-   * Get a specified alert by name. Can return multiple
-   * values if the alert is HOST based.
-   *
-   * @param definitionName alert definition name
-   * @return list of alerts or empty collection
-   */
-  List<Map<String, Object>> getAlert(String definitionName) {
-    getAlerts()[definitionName] ?: []
-  }
-
-  /**
-   * Get a specified alert history by id.
-   *
-   * @param id id of the alert history
-   * @return history or empty collection
-   */
-  Map<String, Object> getAlertHistory(int id) {
-    def response = getAllResources("alert_history/$id")
-    response ? response.AlertHistory : [:]
-  }
-
-  /**
-   * Returns the history of a certain alert.
-   *
-   * @param alertDefinition alert definition name
-   * @param count desired number of result from the latest one
-   * @return list of alert properties or empty collection
-   */
-  def List<Map<String, Object>> getAlertHistory(String alertDefinition, int count) {
-    def result = []
-    def prePredicate = ["AlertHistory/definition_name": alertDefinition]
-    def items = getAllPredictedResources("alert_history", prePredicate).items
-    if (items) {
-      def itemSize = items.size
-      def from = itemSize - count > -1 ? itemSize - count : 0
-      from = from == itemSize ? itemSize - 1 : from
-      items[from..itemSize - 1].collect { it.AlertHistory.id }.each {
-        result << getAlertHistory(it)
-      }
-    }
-    result
+    runServiceCheck('MAPREDUCE2_SERVICE_CHECK', 'MAPREDUCE2')
   }
 
   /**
@@ -481,11 +104,11 @@ class AmbariClient {
    */
   def Map<String, Map<Long, Long>> getDFSSpace() {
     def result = [:]
-    def response = slurp("clusters/${getClusterName()}/services/HDFS/components/NAMENODE", "metrics/dfs")
-    def liveNodes = slurper.parseText(response?.metrics?.dfs?.namenode?.LiveNodes)
+    def response = utils.slurp("clusters/${getClusterName()}/services/HDFS/components/NAMENODE", 'metrics/dfs')
+    def liveNodes = slurper.parseText(response?.metrics?.dfs?.namenode?.LiveNodes as String)
     if (liveNodes) {
       liveNodes.each {
-        if (it.value.adminState == "In Service") {
+        if (it.value.adminState == 'In Service') {
           result << [(it.key.split(':')[0]): [(it.value.remaining as Long): it.value.usedSpace as Long]]
         }
       }
@@ -493,31 +116,8 @@ class AmbariClient {
     result
   }
 
-  /**
-   * Deletes the components from the host.
-   */
-  def deleteHostComponents(String hostName, List<String> components) {
-    components.each {
-      ambari.delete(path: "clusters/${getClusterName()}/hosts/$hostName/host_components/$it")
-    }
-  }
-
-  /**
-   * Deletes the host from the cluster.
-   * Note: Deleting a host from a cluster does not mean it is also
-   * deleted/unregistered from Ambari. It will remain there with UNKNOWN state.
-   */
-  def deleteHost(String hostName) {
-    ambari.delete(path: "clusters/${getClusterName()}/hosts/$hostName")
-  }
-
-  /**
-   * Deletes the host from Ambari.
-   * Note: Deleting a host from a cluster does not mean it is also
-   * deleted/unregistered from Ambari. It will remain there with UNKNOWN state.
-   */
-  def unregisterHost(String hostName) {
-    ambari.delete(path: "hosts/$hostName")
+  def boolean isComponentPresent(def bpMap, def component) {
+    bpMap?.host_groups?.collectNested { it?.components?.name }?.flatten()?.contains(component)
   }
 
   /**
@@ -530,98 +130,11 @@ class AmbariClient {
    * @param hostGroup host group of the blueprint
    */
   def void addComponentsToHosts(List<String> hostNames, String blueprint, String hostGroup) throws HttpResponseException {
-    def bpMap = getBlueprint(blueprint)
+    def bpMap = utils.getBlueprint(blueprint)
     def components = bpMap?.host_groups?.find { it.name.equals(hostGroup) }?.components?.collect { it.name }
     if (components) {
       addComponentsToHosts(hostNames, components)
     }
-  }
-
-  /**
-   * Adds the given components to the given hosts. It does not install them. To install the components use
-   * the {@link #installAllComponentsOnHosts(java.util.List)}.
-   * Only existing service components can be added.
-   *
-   * @param hostNames hosts to install the component to
-   * @param components components to be installed
-   * @throws HttpResponseException in case the component's service is not installed
-   */
-  def void addComponentsToHosts(List<String> hostNames, List<String> components) throws HttpResponseException {
-    def commaSepHostNames = hostNames.join(',')
-    def addRequest = [
-      "RequestInfo": ["query": "Hosts/host_name.in($commaSepHostNames)"],
-      "Body"       : ["host_components": components.collectAll { ["HostRoles": ["component_name": it]] }]
-    ]
-    ambari.post(path: "clusters/${getClusterName()}/hosts", body: new JsonBuilder(addRequest).toPrettyString(), { it })
-  }
-
-  /**
-   * In case of adding new nodes to the cluster these nodes should be added to the
-   * appropriate config groups as well otherwise they'll use the default one. This method
-   * adds the given hosts to the desired config groups for every config type like: yarn-site, hdfs-site etc..
-   *
-   * @param hosts hosts to be added to the groups
-   * @param tag tag of the config group
-   */
-  def void addHostsToConfigGroups(List<String> hosts, String tag) {
-    def groups = getConfigGroups(tag)
-    if (groups) {
-      groups.each {
-        it.remove("href")
-        def currentHosts = it.ConfigGroup.hosts
-        hosts.each {
-          currentHosts << ["host_name": it]
-        }
-        def Map<String, ?> putRequestMap = [:]
-        putRequestMap.put('requestContentType', ContentType.URLENC)
-        putRequestMap.put('path', "clusters/${getClusterName()}/config_groups/$it.ConfigGroup.id")
-        putRequestMap.put('body', new JsonBuilder(it).toPrettyString());
-        ambari.put(putRequestMap)
-      }
-    }
-  }
-
-  /**
-   * Starts the given components on a host.
-   * To start all components on hosts use the {@link #startAllServices} method.
-   *
-   * @return map of the component names and their request id since its an async call
-   * @throws HttpResponseException in case the component is not found
-   */
-  def Map<String, Integer> startComponentsOnHost(String hostName, List<String> components) throws HttpResponseException {
-    setComponentsState(hostName, components, "STARTED")
-  }
-
-  /**
-   * Stops the given components on a host.
-   *
-   * @return map of the component names and their request id since its an async call
-   * @throws HttpResponseException in case the component is not found
-   */
-  def Map<String, Integer> stopComponentsOnHost(String hostName, List<String> components) throws HttpResponseException {
-    setComponentsState(hostName, components, "INSTALLED")
-  }
-
-  /**
-   * Stops all the components on the specified hosts.
-   *
-   * @param hostNames list of FQDN
-   * @return request id since its an async call
-   * @throws HttpResponseException in case of any rest exception
-   */
-  def int stopAllComponentsOnHosts(List<String> hostNames) throws HttpResponseException {
-    setAllComponentsStateToInstalled(hostNames, "Stop all components on hosts")
-  }
-
-  /**
-   * Installs the added components on the specified hosts.
-   *
-   * @param hostNames list of FQDN
-   * @return request id since its an async call
-   * @throws HttpResponseException in case of any rest exception
-   */
-  def int installAllComponentsOnHosts(List<String> hostNames) throws HttpResponseException {
-    setAllComponentsStateToInstalled(hostNames, "Install all components on hosts")
   }
 
   /**
@@ -634,89 +147,9 @@ class AmbariClient {
    */
   def int startAllComponents(String blueprint, String hostGroup) throws HttpResponseException {
     def categories = getComponentsCategory(blueprint, hostGroup)
-    def components = categories.findAll { it.value.equals("SLAVE") || it.value.equals("MASTER") }.keySet()
+    def components = categories.findAll { it.value.equals('SLAVE') || it.value.equals('MASTER') }.keySet()
     def query = "HostRoles/component_name.in(${components.join(',')})"
-    setAllComponentsState(getClusterName(), "STARTED", "Start all components", query)
-  }
-
-  /**
-   * Checks whether the blueprint exists or not.
-   *
-   * @param id id of the blueprint
-   * @return true if exists false otherwise
-   */
-  def boolean doesBlueprintExist(String id) {
-    def result = false
-    try {
-      def Map resourceRequest = getResourceRequestMap("blueprints/$id", ['fields': "Blueprints"])
-      def jsonResponse = getSlurpedResource(resourceRequest)
-      result = !(jsonResponse.status)
-    } catch (e) {
-      log.info("Blueprint does not exist", e)
-    }
-    return result
-  }
-
-  /**
-   * Checks whether there are available blueprints or not.
-   *
-   * @return true if blueprints are available false otherwise
-   */
-  def boolean isBlueprintAvailable() {
-    return getBlueprints().items?.size > 0
-  }
-
-  /**
-   * Returns a pre-formatted String of the blueprint.
-   *
-   * @param id id of the blueprint
-   * @return formatted String
-   */
-  def String showBlueprint(String id) {
-    def resp = getBlueprint(id)
-    if (resp) {
-      def groups = resp.host_groups.collect {
-        def name = it.name
-        def comps = it.components.collect { it.name.padRight(PAD).padLeft(PAD + 10) }.join("\n")
-        return "HOSTGROUP: $name\n$comps"
-      }.join("\n")
-      return "[${resp.Blueprints.stack_name}:${resp.Blueprints.stack_version}]\n$groups"
-    }
-    return "Not found"
-  }
-
-  /**
-   * Returns the host group - components mapping of the blueprint.
-   *
-   * @param id id of the blueprint
-   * @return Map where the key is the host group and the value is the list of components
-   */
-  def Map<String, List<String>> getBlueprintMap(String id) {
-    def result = getBlueprint(id)?.host_groups?.collectEntries { [(it.name): it.components.collect { it.name }] }
-    result ?: [:]
-  }
-
-  /**
-   * Returns a pre-formatted String of the blueprints.
-   *
-   * @return formatted blueprint list
-   */
-  def String showBlueprints() {
-    getBlueprints().items.collect {
-      "${it.Blueprints.blueprint_name.padRight(PAD)} [${it.Blueprints.stack_name}:${it.Blueprints.stack_version}]"
-    }.join("\n")
-  }
-
-  /**
-   * Returns a Map containing the blueprint name - stack association.
-   *
-   * @return Map where the key is the blueprint's name value is the used stack
-   */
-  def Map<String, String> getBlueprintsMap() {
-    def result = getBlueprints().items?.collectEntries {
-      [(it.Blueprints.blueprint_name): it.Blueprints.stack_name + ":" + it.Blueprints.stack_version]
-    }
-    result ?: new HashMap()
+    setAllComponentsState(getClusterName(), 'STARTED', 'Start all components', query)
   }
 
   /**
@@ -729,7 +162,7 @@ class AmbariClient {
   def Map<String, List<String>> recommendAssignments(String blueprint) throws InvalidHostGroupHostAssociation {
     def result = [:]
     def hostNames = getHostStatuses().keySet() as List
-    def groups = getBlueprint(blueprint)?.host_groups?.collect { ["name": it.name, "cardinality": it.cardinality] }
+    def groups = utils.getBlueprint(blueprint)?.host_groups?.collect { ['name': it.name, 'cardinality': it.cardinality] }
     if (hostNames && groups) {
       def groupSize = groups.size()
       def hostSize = hostNames.size()
@@ -760,384 +193,6 @@ class AmbariClient {
   }
 
   /**
-   * Returns the name of the host groups for a given blueprint.
-   *
-   * @param blueprint id of the blueprint
-   * @return host group list or empty list
-   */
-  def List<String> getHostGroups(String blueprint) {
-    def result = getBlueprint(blueprint)
-    result ? result.host_groups.collect { it.name } : new ArrayList<String>()
-  }
-
-  /**
-   * Returns a pre-formatted String of a blueprint used by the cluster.
-   *
-   * @return formatted String
-   */
-  def String showClusterBlueprint() {
-    ambari.get(path: "clusters/${getClusterName()}", query: ['format': "blueprint"]).data.text
-  }
-
-  /**
-   * Returns the active Stack urls as json
-   *
-   * @param stack name of the stack e.g HDP
-   * @param stackVersion version of the stack e.g. 2.2
-   * @param osType type of os e.g redhat6
-   * @param repoId e.g HDP-2.2 or HDP-UTILS-1.1.0.20
-   * @throws HttpResponseException in case of error
-   */
-  def String getStackRepositoryAsJson(String stack, String stackVersion, String osType, String repoId) throws HttpResponseException {
-    String path = "stacks/$stack/versions/$stackVersion/operating_systems/$osType/repositories/$repoId";
-    Map resourceRequestMap = getResourceRequestMap(path, null)
-    return getRawResource(resourceRequestMap)
-  }
-
-  /**
-   * Adds a Stack Repository to the Ambari server.
-   * This API may be invoked multiple times to set Base URL for multiple OS types or Stack versions. If this step is not performed,
-   * by default, blueprints will use the latest Base URL defined in the Stack. Exception is thrown if fails.
-   *
-   * @param stack name of the stack e.g HDP
-   * @param stackVersion version of the stack e.g. 2.2
-   * @param osType type of os e.g redhat6
-   * @param repoId e.g HDP-2.2 or HDP-UTILS-1.1.0.20
-   * @param baseUrl the url of repo
-   * @param verify verify the base url
-   * @throws HttpResponseException in case of error
-   */
-  def String addStackRepository(String stack, String stackVersion, String osType, String repoId, String baseUrl, boolean verify) throws HttpResponseException {
-    def Map bodyMap = [
-      "Repositories": ["base_url": baseUrl, "verify_base_url": verify]
-    ]
-    def Map<String, ?> putRequestMap = [:]
-    putRequestMap.put('requestContentType', ContentType.URLENC)
-    putRequestMap.put('path', "stacks/$stack/versions/$stackVersion/operating_systems/$osType/repositories/$repoId")
-    putRequestMap.put('body', new JsonBuilder(bodyMap).toPrettyString());
-    ambari.put(putRequestMap)
-  }
-
-  /**
-   * Adds a blueprint to the Ambari server. Exception is thrown if fails.
-   *
-   * @param json blueprint as json
-   * @throws HttpResponseException in case of error
-   */
-  def String addBlueprint(String json) throws HttpResponseException {
-    postBlueprint(json)
-  }
-
-  /**
-   * Extends the given blueprint with global configurations.
-   *
-   * @param blueprintJson original blueprint JSON as String
-   * @param newConfigs global configuration map - <site_config_name <config_name, config_value>>
-   * @return extended blueprints JSON as String
-   */
-  def String extendBlueprintGlobalConfiguration(String blueprintJson, Map<String, Map<String, String>> newConfigs) {
-    def blueprintMap = slurper.parseText(blueprintJson)
-    def configurations = blueprintMap.configurations
-    if (!configurations) {
-      if (newConfigs) {
-        def conf = []
-        newConfigs.each { conf << [(it.key): ["properties": it.value]] }
-        blueprintMap << ["configurations": conf]
-      }
-    } else {
-      newConfigs.each {
-        def site = it.key
-        def index = indexOfConfig(configurations, site)
-        if (index == -1) {
-          configurations << ["$site": ["properties": it.value]]
-        } else {
-          def existingConf = configurations.get(index)
-          existingConf."$site".properties << it.value
-        }
-      }
-    }
-    new JsonBuilder(blueprintMap).toPrettyString()
-  }
-
-  /**
-   * Extends the given blueprint with host group level configurations.
-   *
-   * @param blueprintJson original blueprint JSON as String
-   * @param newConfigs host group level configurations - <host_group_name <site_config_name <config_name, config_value>>>
-   * @return extended blueprints JSON as String
-   */
-  def String extendBlueprintHostGroupConfiguration(String blueprintJson, Map<String, Map<String, Map<String, String>>> newConfigs) {
-    def blueprintMap = slurper.parseText(blueprintJson)
-    for (int j = 0; j < newConfigs.size(); j++) {
-      def configurations = blueprintMap.host_groups.find { it.name == newConfigs.keySet().getAt(j) }.configurations
-      if (!configurations) {
-        if (newConfigs) {
-          def conf = []
-          newConfigs.get(newConfigs.keySet().getAt(j)).each { conf << [(it.key): it.value] }
-          blueprintMap.host_groups.find { it.name == newConfigs.keySet().getAt(j) } << ["configurations": conf]
-        }
-      } else {
-        newConfigs.get(newConfigs.keySet().getAt(j)).each {
-          def site = it.key
-          def index = indexOfConfig(configurations, site)
-          if (index == -1) {
-            configurations << ["$site": it.value]
-          } else {
-            def existingConf = configurations.get(index)
-            existingConf."$site" << it.value
-          }
-        }
-      }
-    }
-    new JsonBuilder(blueprintMap).toPrettyString()
-  }
-
-  /**
-   * Adds the default blueprints.
-   *
-   * @throws HttpResponseException in case of error
-   */
-  def void addDefaultBlueprints() throws HttpResponseException {
-    addBlueprint(getResourceContent("blueprints/multi-node-hdfs-yarn"))
-    addBlueprint(getResourceContent("blueprints/single-node-hdfs-yarn"))
-    addBlueprint(getResourceContent("blueprints/lambda-architecture"))
-    addBlueprint(getResourceContent("blueprints/hdp-singlenode-default"))
-    addBlueprint(getResourceContent("blueprints/hdp-multinode-default"))
-  }
-
-  /**
-   * Creates a cluster with the given blueprint and host group - host association.
-   *
-   * @param clusterName name of the cluster
-   * @param blueprintName blueprint id used to create this cluster
-   * @param hostGroups Map<String, List<String> key - host group, value - host list
-   * @return true if the creation was successful false otherwise
-   * @throws HttpResponseException in case of error
-   */
-  def void createCluster(String clusterName, String blueprintName, Map<String, List<String>> hostGroups) throws HttpResponseException {
-    if (debugEnabled) {
-      println "[DEBUG] POST ${ambari.getUri()}clusters/$clusterName"
-    }
-    ambari.post(path: "clusters/$clusterName", body: createClusterJson(blueprintName, hostGroups), { it })
-  }
-
-  /**
-   * Deletes the cluster.
-   *
-   * @param clusterName name of the cluster
-   * @throws HttpResponseException in case of error
-   */
-  def void deleteCluster(String clusterName) throws HttpResponseException {
-    if (debugEnabled) {
-      println "[DEBUG] DELETE ${ambari.getUri()}clusters/$clusterName"
-    }
-    ambari.delete(path: "clusters/$clusterName")
-  }
-
-  /**
-   * Returns the active cluster as json
-   *
-   * @return cluster as json String
-   * @throws HttpResponseException in case of error
-   */
-  def String getClusterAsJson() throws HttpResponseException {
-    String path = "clusters/" + getClusterName();
-    Map resourceRequestMap = getResourceRequestMap(path, null)
-    return getRawResource(resourceRequestMap)
-  }
-
-  /**
-   * Returns all clusters as json
-   *
-   * @return json String
-   * @throws HttpResponseException in case of error
-   */
-  def getClustersAsJson() throws HttpResponseException {
-    Map resourceRequestMap = getResourceRequestMap("clusters", null)
-    return getRawResource(resourceRequestMap)
-  }
-
-  /**
-   * Modify an existing configuration. Be ware you'll have to provide the whole configuration
-   * otherwise properties might get lost.
-   *
-   * @param type type of the configuration e.g capacity-scheduler
-   * @param properties properties to be used
-   */
-  def modifyConfiguration(String type, Map<String, String> properties) {
-    Map bodyMap = [
-      "Clusters": ["desired_config": ["type": type, "tag": "version${System.currentTimeMillis()}", "properties": properties]]
-    ]
-    def Map<String, ?> putRequestMap = [:]
-    putRequestMap.put('requestContentType', ContentType.URLENC)
-    putRequestMap.put('path', "clusters/${getClusterName()}")
-    putRequestMap.put('body', new JsonBuilder(bodyMap).toPrettyString());
-    ambari.put(putRequestMap)
-  }
-
-  /**
-   * Returns a pre-formatted String of the clusters.
-   *
-   * @return pre-formatted cluster list
-   */
-  def String showClusterList() {
-    getClusters().items.collect {
-      "[$it.Clusters.cluster_id] $it.Clusters.cluster_name:$it.Clusters.version"
-    }.join("\n")
-  }
-
-  /**
-   * Returns a map with <status, list of request ids> based on statuses parameter.
-   * @param statuses the relevant statuses
-   * @return the status map
-   */
-  def Map<String, List<Integer>> getRequests(String... statuses) {
-    def reqs = getAllResources("requests", "Requests/request_status,Requests/id")?.items?.Requests
-    def resp = [:]
-    reqs.each {
-      def reqStatus = it?.request_status
-      if (!statuses || reqStatus in statuses) {
-        def reqlist = resp[reqStatus]
-        if (reqlist == null) {
-          reqlist = []
-          resp[reqStatus] = reqlist
-        }
-        reqlist << it?.id
-      }
-    }
-    return resp
-  }
-
-  /**
-   * Returns the task properties as Map.
-   *
-   * @param request request id; default is 1
-   * @return property Map or empty Map
-   */
-  def getTasks(request = 1) {
-    getAllResources("requests/$request", "tasks/Tasks")
-  }
-
-  /**
-   * Returns the install progress state. If the install failed -1 returned.
-   *
-   * @param request request id; default is 1
-   * @return progress in percentage
-   */
-  def BigDecimal getRequestProgress(request = 1) {
-    def response = getAllResources("requests/$request", "Requests")
-    def String status = response?.Requests?.request_status
-    if (status && status.equals("FAILED")) {
-      return new BigDecimal(-1)
-    }
-    return response?.Requests?.progress_percent
-  }
-
-  /**
-   * Returns a pre-formatted task list.
-   *
-   * @param request request id; default is 1
-   * @return pre-formatted task list
-   */
-  def String showTaskList(request = 1) {
-    getTasks(request)?.tasks.collect { "${it.Tasks.command_detail.padRight(PAD)} [${it.Tasks.status}]" }.join("\n")
-  }
-
-  /**
-   * Returns a Map containing the task's command detail as key and the task's status as value.
-   *
-   * @param request request id; default is 1
-   * @return key task command detail; task value status
-   */
-  def Map<String, String> getTaskMap(request = 1) {
-    def result = getTasks(request).tasks?.collectEntries { [(it.Tasks.command_detail): it.Tasks.status] }
-    result ?: new HashMap()
-  }
-
-  /**
-   * Returns the available host names and their states. It also
-   * contains hosts which are not part of the cluster, but are connected
-   * to ambari.
-   *
-   * @return hostname state association
-   */
-  def Map<String, String> getHostStatuses() {
-    getHosts().items.collectEntries { [(it.Hosts.host_name): it.Hosts.host_status] }
-  }
-
-  def Map<String, String> getHostNames() {
-    getHosts().items.collectEntries { [(it.Hosts.public_host_name): it.Hosts.host_name] }
-  }
-
-  /**
-   * Returns the names of the hosts which have the given state. It also
-   * contains hosts which are not part of the cluster, but are connected
-   * to ambari.
-   */
-  def Map<String, String> getHostNamesByState(String state) {
-    getHostStatuses().findAll { it.value == state }
-  }
-
-  /**
-   * Returns a pre-formatted list of the hosts.
-   *
-   * @return pre-formatted String
-   */
-  def String showHostList() {
-    getHosts().items.collect {
-      "$it.Hosts.host_name [$it.Hosts.host_status] $it.Hosts.ip $it.Hosts.os_type:$it.Hosts.os_arch"
-    }.join("\n")
-  }
-
-  /**
-   * Returns a pre-formatted list of the service components.
-   *
-   * @return pre-formatted String
-   */
-  def String showServiceComponents() {
-    getServices().items.collect {
-      def name = it.ServiceInfo.service_name
-      def state = it.ServiceInfo.state
-      def componentList = getServiceComponents(name).items.collect {
-        "    ${it.ServiceComponentInfo.component_name.padRight(PAD)}  [$it.ServiceComponentInfo.state]"
-      }.join("\n")
-      "${name.padRight(PAD)} [$state]\n$componentList"
-    }.join("\n")
-  }
-
-  /**
-   * Returns the service components properties as Map where the key is the service name and
-   * value is a component - state association.
-   *
-   * @return service name - [component name - status]
-   */
-  def Map<String, Map<String, String>> getServiceComponentsMap() {
-    def result = getServices().items.collectEntries {
-      def name = it.ServiceInfo.service_name
-      def componentList = getServiceComponents(name).items.collectEntries {
-        [(it.ServiceComponentInfo.component_name): it.ServiceComponentInfo.state]
-      }
-      [(name): componentList]
-    }
-    result ?: new HashMap()
-  }
-
-  /**
-   * Returns all the components states grouped by hosts.
-   *
-   * @return host name - [component name - state]
-   */
-  def Map<String, Map<String, String>> getHostComponentsStates() {
-    getAllResources("hosts", "host_components/HostRoles/state").items.collectEntries {
-      def hostName = it.Hosts.host_name
-      def components = it.host_components.collectEntries {
-        [(it.HostRoles.component_name): it.HostRoles.state]
-      }
-      [(hostName): components]
-    }
-  }
-
-  /**
    * Returns the type of the components of a given host group in a given blueprint.
    * There are 3 types: SLAVE, CLIENT, MASTER
    *
@@ -1146,7 +201,7 @@ class AmbariClient {
    * @return map where the key is the component's name and the value is the category
    */
   def Map<String, String> getComponentsCategory(String blueprintId, String hostGroup) {
-    def bpMap = getBlueprint(blueprintId)
+    def bpMap = utils.getBlueprint(blueprintId)
     def components = bpMap?.host_groups?.find { it.name.equals(hostGroup) }?.components?.collect { it.name }
     getComponentsCategory(components)
   }
@@ -1159,7 +214,7 @@ class AmbariClient {
    * @return map where the key is the component's name and the value is the category
    */
   def Map<String, String> getComponentsCategory(String blueprintId) {
-    def bpMap = getBlueprint(blueprintId)
+    def bpMap = utils.getBlueprint(blueprintId)
     def components = bpMap?.host_groups?.components?.name?.flatten()
     getComponentsCategory(components)
   }
@@ -1174,7 +229,7 @@ class AmbariClient {
   def Map<String, String> getComponentsCategory(List<String> components) {
     def result = [:]
     components.each {
-      def json = slurp("clusters/${getClusterName()}/components/$it", "ServiceComponentInfo")
+      def json = utils.slurp("clusters/${getClusterName()}/components/$it", 'ServiceComponentInfo')
       def category = json?.ServiceComponentInfo?.category
       if (category) {
         result << [(it): category]
@@ -1189,161 +244,7 @@ class AmbariClient {
    * @return status
    */
   def String healthCheck() {
-    ambari.get(path: "check", headers: ["Accept": ContentType.TEXT]).data.text
-  }
-
-  /**
-   * Returns a pre-formatted service list.
-   *
-   * @return formatted String
-   */
-  def String showServiceList() {
-    getServices().items.collect { "${it.ServiceInfo.service_name.padRight(PAD)} [$it.ServiceInfo.state]" }.join("\n")
-  }
-
-  /**
-   * Returns the services properties as Map where the key is the service's name and the values is the service's state.
-   *
-   * @return service name - service state association as Map
-   */
-  def Map<String, String> getServicesMap() {
-    def result = getServices().items.collectEntries { [(it.ServiceInfo.service_name): it.ServiceInfo.state] }
-    result ?: new HashMap()
-  }
-
-  /**
-   * Returns a pre-formatted component list of a host.
-   *
-   * @param host which host's components are requested
-   * @return formatted String
-   */
-  def String showHostComponentList(host) {
-    getHostComponents(host).items.collect {
-      "${it.HostRoles.component_name.padRight(PAD)} [$it.HostRoles.state]"
-    }.join("\n")
-  }
-
-  /**
-   * Returns the host's components as Map where the key is the component's name and values is its state.
-   *
-   * @param host which host's components are requested
-   * @return component name - state association
-   */
-  def Map<String, String> getHostComponentsMap(host) {
-    def result = getHostComponents(host)?.items?.collectEntries { [(it.HostRoles.component_name): it.HostRoles.state] }
-    result ?: [:]
-  }
-
-  /**
-   * Returns the blueprint json as String.
-   *
-   * @param id id of the blueprint
-   * @return json as String, exception if thrown is it fails
-   */
-  def String getBlueprintAsJson(id) {
-    Map resourceRequestMap = getResourceRequestMap("blueprints/$id", ['fields': "host_groups,Blueprints"])
-    return getRawResource(resourceRequestMap)
-  }
-
-  /**
-   * Returns a map with service configurations of the 'default' config group.
-   * The keys are the config-types, values are maps with <propertyName, propertyValue> entries
-   *
-   * @return a Map with entries of format <config-type, Map<property, value>>
-   */
-  def Map<String, Map<String, String>> getServiceConfigMap(String type = "") {
-    def configs = getServiceConfigMapByHostGroup(null)
-    type ? [(type): configs.get(type)] : configs
-  }
-
-  /**
-   * Returns the currently active host group level configurations combined with the 'default' config group.
-   *
-   * @param hostGroup name of the host group, in Ambari the config group's name is the same
-   * @return return the configurations in the form of <config-type, <property_key, property_value>>
-   */
-  def Map<String, Map<String, String>> getServiceConfigMapByHostGroup(String hostGroup) {
-    Map<String, Map<String, String>> result = new HashMap<>()
-    def rawConfigs = getAllPredictedResources("configurations/service_config_versions",
-      ['is_current': "true", 'fields': "*"]).items
-    rawConfigs.each {
-      if (it.group_name == "default" || (hostGroup && it.group_name.contains(hostGroup))) {
-        it.configurations.each {
-          def type = it.type
-          Map props = result.get(type)
-          if (props) {
-            props.putAll(it.properties)
-          } else {
-            result << [(it.type): it.properties]
-          }
-        }
-      }
-    }
-    result
-  }
-
-  /**
-   * Create a default kerberos config (same as with the wizzard)
-   *
-   * @param kdcHost key distribution center's host
-   * @param realm kerberos realm
-   * @param domain kerberos domain
-   */
-  def void createKerberosConfig(String kdcHost, String realm, String domain) {
-    def model = [
-      "KDC_HOST"     : kdcHost,
-      "REALM"        : realm,
-      "DOMAIN"       : domain,
-      "TAG"          : "version${System.currentTimeMillis()}",
-      "ATTR_TEMPLATE": getResourceContent("templates/kerberos-attr-template"),
-      "CONTENT"      : getResourceContent("templates/kerberos-content-template"),]
-    def Map<String, ?> putRequestMap = [:]
-    putRequestMap.put('requestContentType', ContentType.URLENC)
-    putRequestMap.put('path', "clusters/${getClusterName()}")
-    putRequestMap.put('body', createJson("kerberos-config.json", model));
-    ambari.put(putRequestMap)
-  }
-
-  /**
-   * Creates a default kerberos descriptor for the different Hadoop services.
-   *
-   * @param realm kerberos realm
-   */
-  def void createKerberosDescriptor(String realm) {
-    def json = getResourceContent("templates/kerberos-descriptor").replaceAll("actual-realm", realm)
-    ambari.post(path: "clusters/${getClusterName()}/artifacts/kerberos_descriptor", body: json, { it })
-  }
-
-  /**
-   * In order to add/remove or change any service or host related config in a kerberos secure cluster you need
-   * to set the session attributes which contains an admin principal.
-   *
-   * @param principal admin principal
-   * @param password password for the admin principal
-   */
-  def void setKerberosSession(String principal, String password) {
-    def session = ["session_attributes": ["kerberos_admin": ["principal": principal, "password": password]]]
-    def Map<String, ?> kdcPut = [:]
-    kdcPut.put('requestContentType', ContentType.URLENC)
-    kdcPut.put('path', "clusters/${getClusterName()}")
-    kdcPut.put('body', new JsonBuilder(session).toPrettyString());
-    ambari.put(kdcPut)
-  }
-
-  /**
-   * Enables kerberos security on the cluster. It will generate the necessary keytabs using the previously
-   * provided krb5-conf.
-   *
-   * @return id of the request
-   */
-  def int enableKerberos() {
-    def Map<String, ?> putRequestMap = [:]
-    putRequestMap.put('requestContentType', ContentType.URLENC)
-    putRequestMap.put('path', "clusters/${getClusterName()}")
-    putRequestMap.put('body', new JsonBuilder(["Clusters": ["security_type": "KERBEROS"]]).toPrettyString())
-
-    def response = ambari.put(putRequestMap)
-    slurper.parseText(response.getAt("responseData")?.getAt("str"))?.Requests?.id
+    ambari.get(path: 'check', headers: ['Accept': ContentType.TEXT]).data.text
   }
 
   /**
@@ -1356,469 +257,12 @@ class AmbariClient {
     def Map<String, ?> putRequestMap = [:]
     putRequestMap.put('requestContentType', ContentType.URLENC)
     putRequestMap.put('path', "clusters/${getClusterName()}")
-    putRequestMap.put('body', new JsonBuilder(["Clusters": ["security_type": "KERBEROS"]]).toPrettyString())
+    putRequestMap.put('body', new JsonBuilder(['Clusters': ['security_type': 'KERBEROS']]).toPrettyString())
     if (missingOnly) {
       putRequestMap.put('query', ['regenerate_keytabs': 'missing'])
     }
 
     def response = ambari.put(putRequestMap)
-    slurper.parseText(response.getAt("responseData")?.getAt("str"))?.Requests?.id
+    slurper.parseText(response.getAt('responseData')?.getAt('str') as String)?.Requests?.id
   }
-
-  /**
-   * Add a service to the cluster.
-   *
-   * @param serviceName name of the service
-   */
-  def void addService(String serviceName) {
-    ambari.post(path: "clusters/${getClusterName()}/services",
-      body: new JsonBuilder(["ServiceInfo": ["service_name": serviceName]]).toPrettyString(), { it })
-  }
-
-  /**
-   * Add a service component to the cluster.
-   *
-   * @param serviceName name of the service
-   * @param component component name
-   */
-  def void addServiceComponent(String serviceName, String component) {
-    def body = ["components": [["ServiceComponentInfo": ["component_name": component]]]]
-    def Map<String, ?> requestMap = [:]
-    requestMap.put('path', "clusters/${getClusterName()}/services")
-    requestMap.put('query', ['ServiceInfo/service_name': serviceName])
-    requestMap.put('body', new JsonBuilder(body).toPrettyString());
-    ambari.post(requestMap, { it })
-  }
-
-  /**
-   * Sets the given service's state to the desired value.
-   *
-   * @param serviceName name of the service
-   * @param state desired state
-   * @return id of the request
-   */
-  def int setServiceState(String serviceName, String state) {
-    def body = ["RequestInfo": ["context"        : "Set service: $serviceName state to: $state",
-                                "operation_level": ["level": "CLUSTER", "cluster_name": getClusterName()]],
-                "Body"       : ["ServiceInfo": ["state": state]]]
-    def Map<String, ?> requestMap = [:]
-    requestMap.put('requestContentType', ContentType.URLENC)
-    requestMap.put('path', "clusters/${getClusterName()}/services")
-    requestMap.put('query', ['ServiceInfo/service_name': serviceName, 'ServiceInfo/state': state])
-    requestMap.put('body', new JsonBuilder(body).toPrettyString());
-
-    def response = ambari.put(requestMap)
-    slurper.parseText(response.getAt("responseData")?.getAt("str"))?.Requests?.id
-  }
-
-  /**
-   * Starts all the services.
-   *
-   * @return id of the request since its an async call
-   */
-  def int startAllServices() {
-    log.debug("Starting all services ...")
-    manageService("Start All Services", "STARTED")
-  }
-
-  /**
-   * Stops all the services.
-   *
-   * @return id of the request since its an async call
-   */
-  def int stopAllServices() {
-    log.debug("Stopping all services ...")
-    manageService("Stop All Services", "INSTALLED")
-  }
-
-  /**
-   * Starts the given service.
-   *
-   * @param service name of the service
-   * @return id of the request
-   */
-  def int startService(String service) {
-    manageService("Starting $service", "STARTED", service)
-  }
-
-  /**
-   * Stops the given service.
-   *
-   * @param service name of the service
-   * @return id of the request
-   */
-  def int stopService(String service) {
-    manageService("Stopping $service", "INSTALLED", service)
-  }
-
-  def boolean servicesStarted() {
-    return servicesStatus(true)
-  }
-
-  def boolean servicesStopped() {
-    return servicesStatus(false)
-  }
-
-  /**
-   * Returns the public hostnames of the hosts which the host components are installed to.
-   */
-  def List<String> getPublicHostNames(String hostComponent) {
-    def hosts = getHostNamesByComponent(hostComponent)
-    if (hosts) {
-      return hosts.collect() { resolveInternalHostName(it) }
-    } else {
-      return []
-    }
-  }
-
-  /**
-   * Returns the internal host names of the hosts which the host components are installed to.
-   */
-  def List<String> getHostNamesByComponent(String component) {
-    def hostRoles = getAllResources("host_components", "HostRoles/component_name")
-    def hosts = hostRoles?.items?.findAll { it.HostRoles.component_name.equals(component) }?.HostRoles?.host_name
-    hosts ?: []
-  }
-
-  /**
-   * Restarts the given components of a service.
-   */
-  def int restartServiceComponents(String service, List<String> components) {
-    def filter = components.collect {
-      ["service_name": service, "component_name": it, "hosts": getHostNamesByComponent(it).join(",")]
-    }
-    Map bodyMap = [
-      "RequestInfo"              : [command: "RESTART", context: "Restart $service components $components"],
-      "Requests/resource_filters": filter
-    ]
-    ambari.post(path: "clusters/${getClusterName()}/requests", body: new JsonBuilder(bodyMap).toPrettyString(), {
-      getRequestId(it)
-    })
-  }
-
-  /**
-   * Returns the names of the hosts which are in the cluster.
-   */
-  def List<String> getClusterHosts() {
-    slurp("clusters/${getClusterName()}")?.hosts?.Hosts?.host_name
-  }
-
-  /**
-   * Resolves an internal hostname to a public one.
-   */
-  def String resolveInternalHostName(String internalHostName) {
-    slurp("clusters/${getClusterName()}/hosts/$internalHostName")?.Hosts?.public_host_name
-  }
-
-  def private boolean servicesStatus(boolean starting) {
-    def String status = (starting) ? "STARTED" : "INSTALLED"
-    Map serviceComponents = getServicesMap();
-    boolean allInState = true;
-    serviceComponents.values().each { val ->
-      log.debug("Service: {}", val)
-      allInState = allInState && val.equals(status)
-    }
-    return allInState;
-  }
-
-  def private manageService(String context, String state, String service = "") {
-    Map bodyMap = [
-      RequestInfo: [context: context],
-      ServiceInfo: [state: state]
-    ]
-    JsonBuilder builder = new JsonBuilder(bodyMap)
-    def path = "${ambari.getUri()}clusters/${getClusterName()}/services"
-    if (service) {
-      path += "/$service"
-    }
-    def Map<String, ?> putRequestMap = [:]
-    putRequestMap.put('requestContentType', ContentType.URLENC)
-    putRequestMap.put('path', path)
-    putRequestMap.put('query', ['params/run_smoke_test': 'false'])
-    putRequestMap.put('body', builder.toPrettyString());
-
-    def reponse = ambari.put(putRequestMap)
-    slurper.parseText(reponse.getAt("responseData")?.getAt("str"))?.Requests?.id
-  }
-
-  private Map<String, ?> getResourceRequestMap(String path, Map<String, String> queryParams) {
-    def Map requestMap = [:]
-    if (queryParams) {
-      requestMap = ['path': "${ambari.getUri()}" + path, 'query': queryParams]
-    } else {
-      requestMap = ['path': "${ambari.getUri()}" + path]
-    }
-    return requestMap
-  }
-
-  /**
-   * Gets the resource as a text as it;s returned by the server.
-   *
-   * @param resourceRequestMap
-   */
-  private getRawResource(Map resourceRequestMap) {
-    def rawResource = null;
-    try {
-      if (debugEnabled) {
-        println "[DEBUG] GET ${resourceRequestMap.get('path')}"
-      }
-      rawResource = ambari.get(resourceRequestMap)?.data?.text
-    } catch (e) {
-      def clazz = e.class
-      log.error("Error occurred during GET request to {}, exception: ", resourceRequestMap.get('path'), e)
-      if (clazz == NoHttpResponseException.class || clazz == ConnectException.class
-        || clazz == ClientProtocolException.class || clazz == NoRouteToHostException.class
-        || clazz == UnknownHostException.class || (clazz == HttpResponseException.class && e.message == "Bad credentials")) {
-        throw new AmbariConnectionException("Cannot connect to Ambari ${ambari.getUri()}")
-      }
-    }
-    return rawResource
-  }
-
-  /**
-   * Slurps the response text.
-   *
-   * @param resourceRequestMap a map wrapping the resource request components
-   * @return an Object as it's created by the JsonSlurper
-   */
-  private getSlurpedResource(Map resourceRequestMap) {
-    def rawResource = getRawResource(resourceRequestMap)
-    def slurpedResource = (rawResource != null) ? slurper.parseText(rawResource) : rawResource
-    return slurpedResource
-  }
-
-  private def getAllResources(resourceName, fields = "") {
-    slurp("clusters/${getClusterName()}/$resourceName", fields ? "$fields/*" : "")
-  }
-
-  private def getAllPredictedResources(resourceName, predicate = [:]) {
-    slurpPredicate("clusters/${getClusterName()}/$resourceName", predicate)
-  }
-
-  /**
-   * Posts the blueprint JSON to Ambari with name 'bp' in the URL
-   * because it does not matter here. The blueprint's name is
-   * provided in the JSON.
-   *
-   * @param blueprint json
-   * @return response message
-   */
-  private void postBlueprint(String blueprint) {
-    if (debugEnabled) {
-      println "[DEBUG] POST ${ambari.getUri()}blueprints/bp"
-    }
-    ambari.post(path: "blueprints/bp", body: blueprint, { it })
-  }
-
-  private def createClusterJson(String name, Map hostGroups) {
-    def builder = new JsonBuilder()
-    def groups = hostGroups.collect {
-      def hostList = it.value.collect { ['fqdn': it] }
-      [name: it.key, hosts: hostList]
-    }
-    builder { "blueprint" name; "default_password" "admin"; "host_groups" groups }
-    builder.toPrettyString()
-  }
-
-  private def slurp(path, fields = "") {
-
-    def fieldsMap = fields ? ['fields': fields] : [:]
-    def Map resourceReqMap = getResourceRequestMap(path, fieldsMap)
-    def result = getSlurpedResource(resourceReqMap)
-
-    return result
-  }
-
-  private def slurpPredicate(path, predicate) {
-    def Map resourceReqMap = getResourceRequestMap(path, predicate)
-    getSlurpedResource(resourceReqMap)
-  }
-
-  /**
-   * Return the blueprint's properties as a Map.
-   *
-   * @param id id of the blueprint
-   * @return properties as Map
-   */
-  private def getBlueprint(id) {
-    slurp("blueprints/$id", "host_groups,Blueprints")
-  }
-
-  /**
-   * Returns a Map containing the blueprint's properties parsed from the Ambari response json.
-   *
-   * @return blueprint's properties as Map or empty Map
-   */
-  private def getBlueprints() {
-    slurp("blueprints", "Blueprints")
-  }
-
-  /**
-   * Returns a Map containing the cluster's properties parsed from the Ambari response json.
-   *
-   * @return cluster's properties as Map or empty Map
-   */
-  private def getClusters() {
-    slurp("clusters", "Clusters")
-  }
-
-  /**
-   * Returns the available hosts properties as a Map.
-   *
-   * @return Map containing the hosts properties
-   */
-  private def getHosts() {
-    slurp("hosts", "Hosts")
-  }
-
-  /**
-   * Returns the service components properties as Map.
-   *
-   * @param service id of the service
-   * @return service component properties as Map
-   */
-  private def getServiceComponents(service) {
-    getAllResources("services/$service/components", "ServiceComponentInfo")
-  }
-
-  /**
-   * Returns the services properties as Map parsed from Ambari response json.
-   *
-   * @return service properties as Map
-   */
-  private def getServices() {
-    getAllResources("services", "ServiceInfo")
-  }
-
-  private def addComponentToHost(String hostName, String component) {
-    if (debugEnabled) {
-      println "[DEBUG] POST ${ambari.getUri()}clusters/${getClusterName()}/hosts/$hostName/host_components"
-    }
-    ambari.post(path: "clusters/${getClusterName()}/hosts/$hostName/host_components/${component.toUpperCase()}", { it })
-  }
-
-  private def Map<String, Integer> setComponentsState(String hostName, List<String> components, String state)
-    throws HttpResponseException {
-    def resp = [:]
-    components.each {
-      def id = setComponentState(hostName, it, state)
-      if (id) {
-        resp << [(it): id]
-      }
-    }
-    return resp
-  }
-
-  private def setComponentState(String hostName, String component, String state) {
-    if (debugEnabled) {
-      println "[DEBUG] PUT ${ambari.getUri()}clusters/${getClusterName()}/hosts/$hostName/host_components/$component"
-    }
-    Map bodyMap = [
-      HostRoles  : [state: state.toUpperCase()],
-      RequestInfo: [context: "${component.toUpperCase()} ${state.toUpperCase()}"]
-    ]
-    def Map<String, ?> putRequestMap = [:]
-    putRequestMap.put('requestContentType', ContentType.URLENC)
-    putRequestMap.put('path', "clusters/${getClusterName()}/hosts/$hostName/host_components/${component.toUpperCase()}")
-    putRequestMap.put('body', new JsonBuilder(bodyMap).toPrettyString());
-    def response = ambari.put(putRequestMap).getAt("responseData")?.getAt("str")
-    if (response) {
-      slurper.parseText(response)?.Requests?.id
-    }
-  }
-
-  /**
-   * Sets the state of all components to the desired state.
-   *
-   * @param clusterName name of the cluster
-   * @param state desired state
-   * @param context context shown on the UI
-   * @return id of the request
-   */
-  private def int setAllComponentsState(clusterName, state, context, query) throws HttpResponseException {
-    def reqInfo = [
-      "context"        : context,
-      "operation_level": ["level": "HOST_COMPONENT", "cluster_name": clusterName],
-      "query"          : query
-    ]
-    def reqBody = ["HostRoles": ["state": state]]
-    def Map<String, ?> putRequestMap = [:]
-    putRequestMap.put('requestContentType', ContentType.URLENC)
-    putRequestMap.put('path', "clusters/$clusterName/host_components")
-    putRequestMap.put('body', new JsonBuilder(["RequestInfo": reqInfo, "Body": reqBody]).toPrettyString())
-    if (state == "INSTALLED") {
-      putRequestMap.put('query', ['state': 'INIT'])
-    }
-    def response = ambari.put(putRequestMap).getAt("responseData")?.getAt("str")
-    response ? slurper.parseText(response)?.Requests?.id : -1
-  }
-
-  /**
-   * Returns the properties of the host components as a Map parsed from the Ambari response json.
-   *
-   * @param host which host's components are requested
-   * @return component properties as Map
-   */
-  private def getHostComponents(host) {
-    getAllResources("hosts/$host/host_components", "HostRoles")
-  }
-
-  private String getResourceContent(name) {
-    getClass().getClassLoader().getResourceAsStream(name)?.text
-  }
-
-  private int indexOfConfig(List<Map> configurations, String site) {
-    def index = 0
-    for (Map conf : configurations) {
-      if (conf.containsKey(site)) {
-        return index;
-      }
-      index++
-    }
-    return -1;
-  }
-
-  private def int getRequestId(def responseDecorator) {
-    def resp = IOUtils.toString(new InputStreamReader(responseDecorator.entity.content))
-    slurper.parseText(resp)?.Requests?.id
-  }
-
-  private def boolean isComponentPresent(def bpMap, def component) {
-    bpMap?.host_groups?.collectAll { it?.components?.name }.flatten().contains(component)
-  }
-
-  private def String createJson(String templateName, Map bindings) throws Exception {
-    def InputStream inPut = this.getClass().getClassLoader().getResourceAsStream("templates/$templateName");
-    templateProcessor.createTemplate(new InputStreamReader(inPut)).make(bindings);
-  }
-
-  private def getConfigGroups(String tag) {
-    def groups = getAllResources("config_groups", "ConfigGroup")
-    groups.items.findAll {
-      it.ConfigGroup.group_name.endsWith(tag)
-    }
-  }
-
-  private def int setAllComponentsStateToInstalled(List<String> hostNames, String context) {
-    setAllComponentsState(getClusterName(), "INSTALLED", context, "HostRoles/host_name.in(${hostNames.join(',')})")
-  }
-
-  private SSLContext setupSSLContext(clientCertPath, clientKeyPath, serverCertPath) {
-    Security.addProvider(new org.bouncycastle.jce.provider.BouncyCastleProvider());
-    SSLContext context = SSLContexts.custom()
-      .loadTrustMaterial(KeystoreUtils.createTrustStore(serverCertPath))
-      .loadKeyMaterial(KeystoreUtils.createKeyStore(clientCertPath, clientKeyPath), "consul".toCharArray())
-      .build();
-    return context;
-  }
-
-  private Registry<ConnectionSocketFactory> setupSchemeRegistry(SSLContext sslContext) {
-    RegistryBuilder<ConnectionSocketFactory> registryBuilder = RegistryBuilder.create();
-    registryBuilder.register("http", PlainConnectionSocketFactory.getSocketFactory());
-    if (sslContext != null) {
-      registryBuilder.register("https", new SSLConnectionSocketFactory(sslContext));
-    }
-    return registryBuilder.build();
-  }
-
 }
