+++ src/main/groovy/com/sequenceiq/ambari/client/services/TaskService.groovy
@@ -1,0 +1,66 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.sequenceiq.ambari.client.services
+
+import groovy.util.logging.Slf4j
+
+@Slf4j
+trait TaskService extends ClusterService {
+
+  def getTasks() {
+    return getTasks(1)
+  }
+
+  /**
+   * Returns the task properties as Map.
+   *
+   * @param request request id; default is 1
+   * @return property Map or empty Map
+   */
+  def getTasks(request) {
+    utils.getAllResources("requests/$request", 'tasks/Tasks')
+  }
+
+  def String showTaskList() {
+    return showTaskList(1)
+  }
+  /**
+   * Returns a pre-formatted task list.
+   *
+   * @param request request id; default is 1
+   * @return pre-formatted task list
+   */
+  def String showTaskList(request) {
+    getTasks(request)?.tasks.collect { "${it.Tasks.command_detail.padRight(PAD)} [${it.Tasks.status}]" }.join('\n')
+  }
+
+  def Map<String, String> getTaskMap() {
+    return getTaskMap(1)
+  }
+
+  /**
+   * Returns a Map containing the task's command detail as key and the task's status as value.
+   *
+   * @param request request id; default is 1
+   * @return key task command detail; task value status
+   */
+  def Map<String, String> getTaskMap(request) {
+    def result = getTasks(request).tasks?.collectEntries { [(it.Tasks.command_detail): it.Tasks.status] }
+    result ?: new HashMap()
+  }
+}
